test: Make index of help center with FlexSearch.
NOT A REAL COMMIT MESSAGE - PROGRESS REPORT

I have used render_markdown_path to get the html for the docs,
beautiful soup to remove the html tags, and saved them to a
temporary directory.

I have added flexearch to the node directory and bumped the
version number.

I have written a node.js script in a separate directory to
test FlexSearch and built an index on some help documentation
that I scraped from the web.

I need to write a node.js script to get the html content / text
and write it to a FlexSearch index and export / save the index
for the front end to use.