user-status: Remove deprecated `away` parameter/field.
Removes support for the `away` field in user status objects and
events, which in b2e0b51 became a deprecated way for clients to
access a user's presence status.

Updates the API documentation and changelog for this change. Also
takes advantage of updating the changelog to fix a few small copy
and paste errors in other feature level entries.