click-handlers: Check message group for narrowing via recipient bar.
Previously, when the topic or direct message recipient was clicked on
in the message header bar, the `target_id` for the narrow would be
the ID of the nearest message to the header bar unless the selected
message shared the same recipient. In that case, it would be the
selected message ID.

This behavior could be confusing in search-like views where messages
don't share message header bars or in interleaved views where a
conversation might have multiple message header bars that could be
clicked on.

Instead, we check to see if the nearest and selected messages are in
the same message group, and therefore share the same message header
bar. If they do, then we use the selected message ID. And if they
don't, then we use the nearest message ID.