registration: Use PreregistrationRealm.demo_organization field.
We should set whether the new realm is going to be a demo at the
beginning of the new organization registration process, instead
of setting that in the final step of the process. To do so, we
use the new demo_organization field on the PreregistrationRealm
object.

Removes the is_demo_organization field from the RegistrationForm,
and adds a create_demo field to the superclass RealmDetailsForm.
We're setting the field on the superclass because the validation
for the realm_subdomain field, which is set in the superclass,
will change based on the value of create_demo field.

Because demo organization owners will not have to submit an email
address, we bypass the email confirmation step of the registration
proccess similarly to what we do for the first install of a Zulip
server for the initial realm creation process.

Because demo organization owners will need to agree to any Terms
of Service for the server, we do create a PreregistrationRealm and
redirect to the account registration step.