registration: Make email not required for new realm registration.
Demo organizations will not require users to set an initial email
when creating the new realm.

We set the email field on the RealmCreationForm as not required
and override the form's clean method to only allow the email field
to be an empty string in the case that create_demo is true.